+++ cmd/xferdb/main.go
@@ -1,15 +1,15 @@
 package main
 
 import (
-	"log"
+	"fmt"
 	"os"
 
 	"gitea.homelab.local/nextdevops/XferDB/cmd/xferdb/commands"
 )
 
 func main() {
-	if err := commands.RootCmd.Execute(); err != nil {
-		log.Printf("ERROR: %v", err)
+	if err := commands.Execute(); err != nil {
+		fmt.Fprintln(os.Stderr, "error:", err)
 		os.Exit(1)
 	}
 }